+++ src/Images/Color.cpp
@@ -18,7 +18,7 @@ namespace RayTracer::Images {
         this->_a = color[Color::Types::ALPHA];
     }
 
-    Color::Color(float r, float g, float b, float a) : _r(r), _g(g), _b(b), _a(a) { }
+    Color::Color(double r, double g, double b, double a) : _r(r), _g(g), _b(b), _a(a) { }
 
     Color::Color(const Scenes::ISetting &setting) {
         std::unique_ptr<Scenes::ISetting> tmp = setting.get("r");
@@ -33,34 +33,34 @@ namespace RayTracer::Images {
     }
 
     Color Color::operator+(const Color &other) const {
-        float new_r = this->_r + other[Types::RED];
-        float new_g = this->_g + other[Types::GREEN];
-        float new_b = this->_b + other[Types::BLUE];
-        float new_a = this->_a + other[Types::ALPHA];
+        double new_r = this->_r + other[Types::RED];
+        double new_g = this->_g + other[Types::GREEN];
+        double new_b = this->_b + other[Types::BLUE];
+        double new_a = this->_a + other[Types::ALPHA];
         return {new_r, new_g, new_b, new_a};
     }
 
     Color Color::operator-(const Color &other) const {
-        float new_r = this->_r - other[Types::RED];
-        float new_g = this->_g - other[Types::GREEN];
-        float new_b = this->_b - other[Types::BLUE];
-        float new_a = this->_a - other[Types::ALPHA];
+        double new_r = this->_r - other[Types::RED];
+        double new_g = this->_g - other[Types::GREEN];
+        double new_b = this->_b - other[Types::BLUE];
+        double new_a = this->_a - other[Types::ALPHA];
         return {new_r, new_g, new_b, new_a};
     }
 
     Color Color::operator*(const Color &other) const {
-        float new_r = this->_r * other[Types::RED];
-        float new_g = this->_g * other[Types::GREEN];
-        float new_b = this->_b * other[Types::BLUE];
-        float new_a = this->_a * other[Types::ALPHA];
+        double new_r = this->_r * other[Types::RED];
+        double new_g = this->_g * other[Types::GREEN];
+        double new_b = this->_b * other[Types::BLUE];
+        double new_a = this->_a * other[Types::ALPHA];
         return {new_r, new_g, new_b, new_a};
     }
 
     Color Color::operator/(const Color &other) const {
-        float new_r = this->_r / other[Types::RED];
-        float new_g = this->_g / other[Types::GREEN];
-        float new_b = this->_b / other[Types::BLUE];
-        float new_a = this->_a / other[Types::ALPHA];
+        double new_r = this->_r / other[Types::RED];
+        double new_g = this->_g / other[Types::GREEN];
+        double new_b = this->_b / other[Types::BLUE];
+        double new_a = this->_a / other[Types::ALPHA];
         return {new_r, new_g, new_b, new_a};
     }
 
@@ -74,7 +74,7 @@ namespace RayTracer::Images {
         return *this;
     }
 
-    const float &Color::operator[](const Color::Types &type) const {
+    const double &Color::operator[](const Color::Types &type) const {
         switch (type) {
             case Types::RED:
                 return this->_r;
@@ -89,7 +89,7 @@ namespace RayTracer::Images {
         }
     }
 
-    float &Color::operator[](const Color::Types &type) {
+    double &Color::operator[](const Color::Types &type) {
         switch (type) {
             case Types::RED:
                 return this->_r;
@@ -104,7 +104,7 @@ namespace RayTracer::Images {
         }
     }
 
-    void Color::set(const Types &type, float value) {
+    void Color::set(const Types &type, double value) {
         this->_mutex.lock();
         switch (type) {
             case Types::RED:
@@ -127,10 +127,10 @@ namespace RayTracer::Images {
 }
 
 std::ostream& operator<<(std::ostream& os, const RayTracer::Images::Color& color) {
-    float a = color[RayTracer::Images::Color::Types::ALPHA];
-    float r = a * color[RayTracer::Images::Color::Types::RED] / 255;
-    float g = a * color[RayTracer::Images::Color::Types::GREEN] / 255;
-    float b = a * color[RayTracer::Images::Color::Types::BLUE] / 255;
+    double a = color[RayTracer::Images::Color::Types::ALPHA];
+    double r = a * color[RayTracer::Images::Color::Types::RED] / 255;
+    double g = a * color[RayTracer::Images::Color::Types::GREEN] / 255;
+    double b = a * color[RayTracer::Images::Color::Types::BLUE] / 255;
 
     os << r << ' ' << g << ' ' << b;
     return os;
